% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Experiment}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{positive integer for iterations}
}
\value{
Table and barplot for data
}
\description{
This function executes a binomial experiment for
a sample of population 10 with a probability of 0.7
with a default iterations set to 100.
}
\examples{
mybin(200)

}
