% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Interval Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{the number of iterations performed on a given distribution}

\item{x}{the size of the sample}

\item{fun}{the statistic being used for the bootstrap interval}

\item{alpha}{1 minus the confidence interval being found}

\item{cx}{location varible for marking on graph the point value}
}
\value{
histogram of values

invisible list of x, ci, and fun
}
\description{
This function calculates a bootstrap interval given a statistic, confidence interval value, number of interations
}
