% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Central Limit Theorem}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{the number of the sample size in the runif function}

\item{iter}{the number of iterations run for said sample size}
}
\value{
histogram of sum values
}
\description{
This function produces a histogram of the sum values
for a runif function
}
\examples{
myclt(10,10000)

}
