% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylogbin.R
\name{mylogbin}
\alias{mylogbin}
\title{Log Binomial Function}
\usage{
mylogbin(x, param, size)
}
\arguments{
\item{x}{the vector of quantiles}

\item{size}{the number of trials}

\item{prob}{probability of success on each trial}
}
\value{
log likelihood of given experiment
}
\description{
This function returns the log likelihood of a binomial probability
experiment
}
\examples{
x=c(1,3,4,5,3,2)
mylogbin(x,1,1000)

}
