% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Distribution}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{integer for mean of normal distribution}

\item{sigma}{integer for sd of normal distribution}

\item{a}{integer for probability P(x<=a)}
}
\value{
Curve of normal distribution with mean=mu and sd=sigma,
    shaded probility area and probability of P(x<=a)
}
\description{
This function draws a normal distribution curve for
with mean=mu and sd=sigma an xlim of (mu-3*sigma,mu+3*sigma)
also shades region that corresponds with P(x<=a)
}
\examples{
myncurve(2,4,3)

}
