% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{P-Value Calculator}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{This is the t calc value you enter}

\item{xmax}{The max x-axis value}

\item{alpha}{the 1-alpha is the confidence interval being used}
}
\value{
The p-value and q.
}
\description{
This function calculates the p-value for a given t-calc value
and creates a graph that properly displays the given data
}
\examples{
mypvalue(t.test(x1,mu=22)$statistic)

}
