% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.r
\name{myread}
\alias{myread}
\title{Reads csv file into R}
\usage{
myread(csv)
}
\arguments{
\item{csv}{Data file with ".csv" ending}
}
\value{
Loaded data file in the R workspace
}
\description{
This function reads a csv file containing data into R so it can be
utilized in an R script or markdown file.
}
\examples{
dird="C:\\\\Users\\\\Tanner Cabaniss\\\\Desktop\\\\"
myread("DDT.csv")

}
