% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TannerCabaniss.R
\docType{Package}
\name{TannerCabaniss}
\alias{TannerCabaniss}
\title{Personal Functions Package}
\description{
This package contains all the functions that are required to
be added to said package by the curriculum of MATH4753
}
\author{
Tanner Cabaniss \email{tanner.l.cabaniss@ou.edu}
}
