% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{95% Confidence Interval}
\usage{
myci(data)
}
\arguments{
\item{data}{the data vector}

\item{n}{the number of data points in the set}
}
\value{
The 95% confidence interval
}
\description{
This function creates a 95% confidence interval for a given set of data
}
\examples{
myci(5,c(15,10,12,13,14))

}
